+++ frontend/src/components/RequestDialog.vue
@@ -8,17 +8,10 @@
       :actions="false"
       @submit="submitRequestTransaction"
     >
-      <RequestFormInputs v-if="!isTransferInProgress" />
-      <TransferStatus v-else-if="transfer" :transfer="transfer" />
-      <Transition name="expand">
-        <div v-if="transferError" class="mt-7 text-right text-lg text-orange-dark">
-          {{ transferError }}
-        </div>
-      </Transition>
+      <RequestFormInputs />
 
       <Teleport v-if="signer" to="#action-button-portal">
         <FormKit
-          v-if="!isTransferInProgress"
           class="w-72 flex flex-row justify-center bg-green"
           type="submit"
           :disabled="!valid"
@@ -26,15 +19,6 @@
         >
           Transfer funds
         </FormKit>
-
-        <FormKit
-          v-if="isTransferInProgress"
-          input-class="w-72 flex flex-row justify-center bg-green"
-          type="button"
-          :disabled="isNewTransferDisabled"
-          @click="newTransfer"
-          >New Transfer</FormKit
-        >
       </Teleport>
     </FormKit>
   </div>
@@ -44,20 +28,24 @@
 import { FormKitFrameworkContext } from '@formkit/core';
 import { FormKit } from '@formkit/vue';
 import { storeToRefs } from 'pinia';
-import { computed, ref, watch } from 'vue';
+import { computed, reactive, ref, watch } from 'vue';
 
+import { Transfer } from '@/actions/transfers';
 import RequestFormInputs from '@/components/RequestFormInputs.vue';
-import TransferStatus from '@/components/TransferStatus.vue';
 import { useRequestFee } from '@/composables/useRequestFee';
-import { useTransfer } from '@/composables/useTransfer';
 import { useConfiguration } from '@/stores/configuration';
 import { useEthereumProvider } from '@/stores/ethereum-provider';
 import { useTransferHistory } from '@/stores/transfer-history';
+import type { ChainWithTokens } from '@/types/config';
+import type { Chain, Token } from '@/types/data';
 import type { RequestFormResult } from '@/types/form';
+import { TokenAmount } from '@/types/token-amount';
+import { UInt256 } from '@/types/uint-256';
 
 const configuration = useConfiguration();
 const ethereumProvider = useEthereumProvider();
 const { provider, signer, signerAddress, chainId } = storeToRefs(ethereumProvider);
+const transferHistory = useTransferHistory();
 
 const requestForm = ref<FormKitFrameworkContext>();
 
@@ -71,33 +59,73 @@ const submitForm = () => {
   requestForm.value?.node.submit();
 };
 
-const transferHistory = useTransferHistory();
-const { transfer, runTransfer, isTransferInProgress, isNewTransferDisabled } = useTransfer(
-  transferHistory.addTransfer,
-);
-
 const submitRequestTransaction = async (formResult: RequestFormResult) => {
   if (!provider.value || !signer.value) {
     throw new Error('No signer available!');
   }
-  await runTransfer(
-    formResult,
-    signer.value,
-    signerAddress.value,
-    fees.value,
-    configuration.chains,
-  );
+  const sourceConfiguration = configuration.chains[formResult.sourceChainId.value];
+  const sourceChain = parseChainFromConfiguration(sourceConfiguration);
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const sourceToken = parseTokenFromChainConfiguration(
+    sourceConfiguration,
+    formResult.tokenAddress.label,
+  )!;
+  const sourceAmount = TokenAmount.parse(formResult.amount, sourceToken);
+
+  const targetConfiguration = configuration.chains[formResult.targetChainId.value];
+  const targetChain = parseChainFromConfiguration(targetConfiguration);
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const targetToken = parseTokenFromChainConfiguration(
+    targetConfiguration,
+    formResult.tokenAddress.label,
+  )!;
+  const targetAmount = TokenAmount.parse(formResult.amount, targetToken);
+  const validityPeriod = new UInt256('600');
+
+  const transfer = reactive(
+    Transfer.new(
+      sourceChain,
+      sourceAmount,
+      targetChain,
+      targetAmount,
+      formResult.toAddress,
+      validityPeriod,
+      fees.value as TokenAmount,
+    ),
+  ) as Transfer;
+
+  transferHistory.addTransfer(transfer);
+  requestForm.value?.node.reset();
+
+  try {
+    await transfer.execute(signer.value, signerAddress.value);
+  } catch (error) {
+    console.error(error);
+    console.log(transfer);
+  }
 };
+
+function parseChainFromConfiguration(configuration: ChainWithTokens): Chain {
+  return {
+    identifier: configuration.identifier,
+    name: configuration.name,
+    rpcUrl: configuration.rpcUrl,
+    requestManagerAddress: configuration.requestManagerAddress,
+    fillManagerAddress: configuration.fillManagerAddress,
+    explorerTransactionUrl: configuration.explorerTransactionUrl,
+  };
+}
+
+function parseTokenFromChainConfiguration(
+  configuration: ChainWithTokens,
+  tokenName: string,
+): Token | undefined {
+  return configuration.tokens.find((token) => token.symbol === tokenName);
+}
 
 watch(chainId, (_, oldChainId) => {
   if (oldChainId !== -1) {
     location.reload();
   }
 });
-
-const transferError = computed(() => transfer.value?.errorMessage);
-
-const newTransfer = () => {
-  transfer.value = undefined;
-};
 </script>
