+++ frontend/src/components/layout/Tabs.vue
@@ -1,6 +1,6 @@
 <template>
-  <div class="w-full h-full flex flex-col">
-    <div class="w-full h-[6rem] flex">
+  <div class="w-full h-full flex flex-col overflow-hidden">
+    <div class="tab-header w-full flex">
       <div
         v-for="header of headers"
         :key="header.label"
@@ -13,7 +13,7 @@
       </div>
     </div>
 
-    <div class="w-full h-full p-16" data-test="tab-content">
+    <div class="tab-content w-full p-16" data-test="tab-content">
       <KeepAlive>
         <component :is="activeTab.content" v-if="activeTab" :key="activeTab.label" />
       </KeepAlive>
@@ -49,3 +49,15 @@ const headers = computed(() =>
   })),
 );
 </script>
+
+<style lang="scss">
+$header-height: 6rem;
+
+.tab-header {
+  min-height: $header-height;
+}
+
+.tab-content {
+  height: calc(100% - #{$header-height});
+}
+</style>
