+++ frontend/src/stores/configuration.ts
@@ -3,8 +3,8 @@ import { defineStore } from 'pinia';
 import { BeamerConfig, ChainWithTokens } from '@/types/config';
 
 export const useConfiguration = defineStore('configuration', {
-  state: () => ({
-    chains: {} as BeamerConfig['chains'],
+  state: (): BeamerConfig => ({
+    chains: {},
   }),
   actions: {
     setChainConfiguration(id: string, configuration: ChainWithTokens) {
